+++ src/oidc-agent/daemonize.c
@@ -2,6 +2,7 @@
 #include "utils/agentLogger.h"
 #include "utils/printer.h"
 
+#include <assert.h>
 #include <fcntl.h>
 #include <signal.h>
 #include <stdio.h>
@@ -17,7 +18,7 @@ void sig_handler(int signo) {
   exit(signo);
 }
 
-void daemonize() {
+pid_t daemonize() {
   fflush(stdout);  // flush before forking, otherwise the buffered content is
                    // printed multiple times. (Only relevant when stdout is
                    // redirected, tty is line-buffered.)
@@ -37,10 +38,7 @@ void daemonize() {
     agent_log(ALERT, "fork %m");
     exit(EXIT_FAILURE);
   } else if (pid > 0) {
-    printStdout("%s=%d; export %s;\n", OIDC_PID_ENV_NAME, pid,
-                OIDC_PID_ENV_NAME);
-    printStdout("echo Agent pid $%s\n", OIDC_PID_ENV_NAME);
-    exit(EXIT_SUCCESS);
+    return pid;
   }
   if (chdir("/") != 0) {
     agent_log(ERROR, "chdir %m");
@@ -49,7 +47,8 @@ void daemonize() {
   close(STDIN_FILENO);
   close(STDOUT_FILENO);
   close(STDERR_FILENO);
-  open("/dev/null", O_RDONLY);
-  open("/dev/null", O_RDWR);
-  open("/dev/null", O_RDWR);
+  assert(open("/dev/null", O_RDONLY) == STDIN_FILENO);
+  assert(open("/dev/null", O_RDWR) == STDOUT_FILENO);
+  assert(open("/dev/null", O_RDWR) == STDERR_FILENO);
+  return pid;
 }