+++ test/src/utils/json/tc_jsonAddStringValue.h
@@ -1,8 +1,0 @@
-#ifndef TEST_UTILS_JSON_JSONADDSTRINGVALUE_H
-#define TEST_UTILS_JSON_JSONADDSTRINGVALUE_H
-
-#include <check.h>
-
-TCase* test_case_jsonAddStringValue();
-
-#endif  // TEST_UTILS_JSON_JSONADDSTRINGVALUE_H