+++ test/src/utils/json/tc_jsonArrayAddStringValue.h
@@ -1,8 +1,0 @@
-#ifndef TEST_UTILS_JSON_JSONARRAYADDSTRINGVALUE_H
-#define TEST_UTILS_JSON_JSONARRAYADDSTRINGVALUE_H
-
-#include <check.h>
-
-TCase* test_case_jsonArrayAddStringValue();
-
-#endif  // TEST_UTILS_JSON_JSONARRAYADDSTRINGVALUE_H