+++ test/src/utils/json/tc_jsonHasKey.h
@@ -1,8 +1,0 @@
-#ifndef TEST_UTILS_JSON_JSONHASKEY_H
-#define TEST_UTILS_JSON_JSONHASKEY_H
-
-#include <check.h>
-
-TCase* test_case_jsonHasKey();
-
-#endif  // TEST_UTILS_JSON_JSONHASKEY_H