+++ test/src/utils/json/tc_jsonStringHasKey.h
@@ -1,8 +1,0 @@
-#ifndef TEST_UTILS_JSON_JSONSTRINGHASKEY_H
-#define TEST_UTILS_JSON_JSONSTRINGHASKEY_H
-
-#include <check.h>
-
-TCase* test_case_jsonStringHasKey();
-
-#endif  // TEST_UTILS_JSON_JSONSTRINGHASKEY_H