+++ test/src/utils/json/tc_jsonToString.h
@@ -1,8 +1,0 @@
-#ifndef TEST_UTILS_JSON_JSONTOSTRING_H
-#define TEST_UTILS_JSON_JSONTOSTRING_H
-
-#include <check.h>
-
-TCase* test_case_jsonToString();
-
-#endif  // TEST_UTILS_JSON_JSONTOSTRING_H