+++ test/src/utils/json/tc_mergeJSONObjects.h
@@ -1,8 +1,0 @@
-#ifndef TEST_UTILS_JSON_MERGEJSONOBJECTS_H
-#define TEST_UTILS_JSON_MERGEJSONOBJECTS_H
-
-#include <check.h>
-
-TCase* test_case_mergeJSONObjects();
-
-#endif  // TEST_UTILS_JSON_MERGEJSONOBJECTS_H